requirements.txt
aws-cdk-lib==2.167.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-ec2
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-autoscaling
aws-cdk.aws-iam
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-codebuild
cdk-nag
pyyaml
